move logic out of app.jsx, into home.jsx index route
